Version 551

+++ hydrus/client/gui/widgets/ClientGUICommon.py
@@ -786,8 +786,8 @@ class BufferedWindowIcon( BufferedWindow ):
         
         painter.setRenderHint( QG.QPainter.SmoothPixmapTransform, True ) # makes any scaling here due to jank thumbs look good
         
-        x_offset = ( self.width() - self._pixmap.width() ) / 2
-        y_offset = ( self.height() - self._pixmap.height() ) / 2
+        x_offset = int( ( self.width() - self._pixmap.width() ) / 2 )
+        y_offset = int( ( self.height() - self._pixmap.height() ) / 2 )
         
         if isinstance( self._pixmap, QG.QImage ):
